+++ src/ShaderAnalyzer.cs
@@ -960,7 +960,10 @@ namespace d4rkpl4y3r
         private List<(string name, List<string> lines)> outputIncludes = new List<(string name, List<string> lines)>();
         private List<string> pragmaOutput;
         private ParsedShader parsedShader;
-        private int meshToggleCount;
+        private int mergedMeshCount;
+        private int minMeshIndex;
+        private int maxMeshIndex;
+        private int localMeshCount;
         private Dictionary<string, string> staticPropertyValues;
         private Dictionary<string, (string type, List<string> values)> arrayPropertyValues;
         private Dictionary<string, string> animatedPropertyValues;
@@ -976,7 +979,9 @@ namespace d4rkpl4y3r
 
         public static List<(string name, List<string> lines)> Run(ParsedShader source,
             Dictionary<string, string> staticPropertyValues = null,
-            int meshToggleCount = 0,
+            int mergedMeshCount = 0,
+            int minMeshIndex = 0,
+            int maxMeshIndex = 0,
             Dictionary<string, (string type, List<string> values)> arrayPropertyValues = null,
             Dictionary<string, string> texturesToNullCheck = null,
             HashSet<string> texturesToMerge = null,
@@ -991,7 +996,10 @@ namespace d4rkpl4y3r
             Thread.CurrentThread.CurrentUICulture = CultureInfo.InvariantCulture;
             var optimizer = new ShaderOptimizer
             {
-                meshToggleCount = meshToggleCount,
+                mergedMeshCount = mergedMeshCount,
+                minMeshIndex = minMeshIndex,
+                maxMeshIndex = maxMeshIndex,
+                localMeshCount = maxMeshIndex - minMeshIndex + 1,
                 staticPropertyValues = staticPropertyValues ?? new Dictionary<string, string>(),
                 arrayPropertyValues = arrayPropertyValues ?? new Dictionary<string, (string type, List<string> values)>(),
                 parsedShader = source,
@@ -1037,7 +1045,9 @@ namespace d4rkpl4y3r
             foreach (var animatedProperty in animatedPropertyValues)
             {
                 string name = animatedProperty.Key;
-                string value = $"d4rkAvatarOptimizer{name}[d4rkAvatarOptimizer_MeshID]";
+                string value = localMeshCount > 1
+                    ? $"d4rkAvatarOptimizer{name}[d4rkAvatarOptimizer_MeshID]"
+                    : $"d4rkAvatarOptimizer{name}_ArrayIndex{minMeshIndex}";
                 output.Add($"{name} = asuint({value}.x + d4rkAvatarOptimizer_Zero) == 0xFF555555 ? {name} : {value};");
             }
         }
@@ -1171,7 +1181,7 @@ namespace d4rkpl4y3r
             List<string> originalVertexShader = null;
             bool needToPassOnMeshOrMaterialID =
                 arrayPropertyValues.Count > 0
-                || (pass.geometry != null && meshToggleCount > 1)
+                || (pass.geometry != null && mergedMeshCount > 1)
                 || animatedPropertyValues.Count > 0;
             if (needToPassOnMeshOrMaterialID)
             {
@@ -1201,11 +1211,14 @@ namespace d4rkpl4y3r
                 }
             }
             InitializeOutputParameter(funcParams, output, !needToPassOnMeshOrMaterialID);
-            if (meshToggleCount > 1)
+            if (mergedMeshCount > 1)
             {
-                output.Add($"d4rkAvatarOptimizer_MeshID = {inParam.name}.{vertexInUv0Member}.z;");
+                output.Add($"d4rkAvatarOptimizer_MeshID = {inParam.name}.{vertexInUv0Member}.z - {minMeshIndex};");
                 InjectDummyCBufferUsage(nullReturn);
-                output.Add("if (!_IsActiveMesh[d4rkAvatarOptimizer_MeshID]) " + nullReturn);
+                if (localMeshCount > 1)
+                    output.Add($"if (!_IsActiveMesh[d4rkAvatarOptimizer_MeshID]) {nullReturn}");
+                else
+                    output.Add($"if (!_IsActiveMesh{minMeshIndex}) {nullReturn}");
             }
             if (arrayPropertyValues.Count > 0)
             {
@@ -1297,7 +1310,7 @@ namespace d4rkpl4y3r
             var inParam = func.parameters.FirstOrDefault(p => p.isInput && p.arraySize >= 0);
             var wrapperStructs = new List<string>();
             bool usesOutputWrapper = animatedPropertyValues.Count > 0 || arrayPropertyValues.Count > 0;
-            bool usesInputWrapper = usesOutputWrapper || meshToggleCount > 1;
+            bool usesInputWrapper = usesOutputWrapper || mergedMeshCount > 1;
             if (usesInputWrapper)
             {
                 wrapperStructs.Add("struct geometryInputWrapper");
@@ -1341,10 +1354,13 @@ namespace d4rkpl4y3r
                     output.Add(inParam.name + "[" + i + "] = d4rkAvatarOptimizer_inputWrapper[" + i + "].d4rkAvatarOptimizer_geometryInput;");
                 }
             }
-            if (meshToggleCount > 1)
+            if (mergedMeshCount > 1)
             {
                 InjectDummyCBufferUsage("return;");
-                output.Add("if (!_IsActiveMesh[d4rkAvatarOptimizer_MeshID]) return;");
+                if (localMeshCount > 1)
+                    output.Add("if (!_IsActiveMesh[d4rkAvatarOptimizer_MeshID]) return;");
+                else
+                    output.Add($"if (!_IsActiveMesh{minMeshIndex}) return;");
             }
             if (arrayPropertyValues.Count > 0)
             {
@@ -1526,10 +1542,12 @@ namespace d4rkpl4y3r
 
         private void InjectDummyCBufferUsage(string nullReturn)
         {
+            if (localMeshCount <= 1)
+                return;
             output.Add("if (d4rkAvatarOptimizer_Zero)");
             output.Add("{");
             string val = "float d4rkAvatarOptimizer_val = _IsActiveMesh[d4rkAvatarOptimizer_MeshID]";
-            for (int i = 0; i < meshToggleCount; i++)
+            for (int i = minMeshIndex; i <= maxMeshIndex; i++)
             {
                 val += " + _IsActiveMesh" + i;
             }
@@ -1537,7 +1555,7 @@ namespace d4rkpl4y3r
             foreach (var animatedProperty in animatedPropertyValues.Keys)
             {
                 val = $"d4rkAvatarOptimizer_val += d4rkAvatarOptimizer{animatedProperty}[d4rkAvatarOptimizer_MeshID].x";
-                for (int i = 0; i < meshToggleCount; i++)
+                for (int i = minMeshIndex; i <= maxMeshIndex; i++)
                 {
                     val += $" + d4rkAvatarOptimizer{animatedProperty}_ArrayIndex{i}.x";
                 }
@@ -1550,27 +1568,29 @@ namespace d4rkpl4y3r
         private void InjectPropertyArrays(ParsedShader.Pass pass)
         {
             pragmaOutput.Add("#pragma skip_variants DYNAMICLIGHTMAP_ON LIGHTMAP_ON LIGHTMAP_SHADOW_MIXING DIRLIGHTMAP_COMBINED SHADOWS_SHADOWMASK");
-            if (meshToggleCount > 1)
+            if (mergedMeshCount > 1)
             {
                 output.Add("cbuffer CBd4rkAvatarOptimizer");
                 output.Add("{");
-                output.Add("float _IsActiveMesh[" + meshToggleCount + "] : packoffset(c0);");
-                for (int i = 0; i < meshToggleCount; i++)
+                if (localMeshCount > 1)
+                    output.Add("float _IsActiveMesh[" + localMeshCount + "] : packoffset(c0);");
+                for (int i = 0; i < localMeshCount; i++)
                 {
-                    output.Add("float _IsActiveMesh" + i + " : packoffset(c" + i + ");");
+                    output.Add($"float _IsActiveMesh{i + minMeshIndex} : packoffset(c{i});");
                 }
-                int currentPackOffset = meshToggleCount;
+                int currentPackOffset = localMeshCount;
                 foreach (var animatedProperty in animatedPropertyValues)
                 {
                     string name = "d4rkAvatarOptimizer" + animatedProperty.Key;
                     string type = animatedProperty.Value;
                     type = Regex.Replace(type, "^(bool|int|uint)([1-4]?)$", "float$2");
-                    output.Add($"{type} {name}[{meshToggleCount}] : packoffset(c{currentPackOffset});");
-                    for (int i = 0; i < meshToggleCount; i++)
+                    if (localMeshCount > 1)
+                        output.Add($"{type} {name}[{localMeshCount}] : packoffset(c{currentPackOffset});");
+                    for (int i = 0; i < localMeshCount; i++)
                     {
-                        output.Add($"{type} {name}_ArrayIndex{i} : packoffset(c{currentPackOffset + i});");
+                        output.Add($"{type} {name}_ArrayIndex{i + minMeshIndex} : packoffset(c{currentPackOffset + i});");
                     }
-                    currentPackOffset += meshToggleCount;
+                    currentPackOffset += localMeshCount;
                 }
                 output.Add("};");
             }
@@ -1771,7 +1791,7 @@ namespace d4rkpl4y3r
             {
                 DuplicateFunctionWithTextureParameter(source, ref sourceLineIndex);
             }
-            else if ((arrayPropertyValues.Count > 0 || meshToggleCount > 1) && line.StartsWith("struct "))
+            else if ((arrayPropertyValues.Count > 0 || mergedMeshCount > 1) && line.StartsWith("struct "))
             {
                 output.Add(line);
                 var match = Regex.Match(line, @"struct\s+(\w+)");
@@ -1859,7 +1879,7 @@ namespace d4rkpl4y3r
                 }
                 else if (line.StartsWith("#pragma"))
                 {
-                    if (((pass.geometry != null && meshToggleCount > 1) || arrayPropertyValues.Count > 0 || animatedPropertyValues.Count > 0)
+                    if (((pass.geometry != null && mergedMeshCount > 1) || arrayPropertyValues.Count > 0 || animatedPropertyValues.Count > 0)
                         &&  Regex.IsMatch(line, @"^#pragma\s+vertex\s+\w+"))
                     {
                         pragmaOutput.Add("#pragma vertex d4rkAvatarOptimizer_vertexWithWrapper");
@@ -1915,7 +1935,7 @@ namespace d4rkpl4y3r
                     output.Add(line);
                     propertyBlockStartParseIndex = lineIndex;
                     propertyBlockInsertionIndex = output.Count;
-                    for (int i = 0; meshToggleCount > 1 && i < meshToggleCount; i++)
+                    for (int i = minMeshIndex; mergedMeshCount > 1 && i <= maxMeshIndex; i++)
                     {
                         propertyBlock.Add("_IsActiveMesh" + i + "(\"Generated Mesh Toggle " + i + "\", Float) = 1");
                     }
@@ -1929,7 +1949,7 @@ namespace d4rkpl4y3r
                             defaultValue = "(0,0,0,0)";
                             type = "Vector";
                         }
-                        for (int i = 0; i < meshToggleCount; i++)
+                        for (int i = minMeshIndex; i <= maxMeshIndex; i++)
                         {
                             propertyBlock.Add($"d4rkAvatarOptimizer{prop.name}_ArrayIndex{i}(\"{prop.name} {i}\", {type}) = {defaultValue}");
                         }