+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### Features
 * Add debug view to show all animated material property paths.
 * Don't inject animated property arrays for merged material properties where none of the original meshes had that property animated.
+* Calculate min and max mesh id for each merged material and only add animated property arrays for that range.
 
 ### Bug Fixes
 * Delete shader properties that start with _ShaderOptimizer from the optimized shader. This is to prevent Kaj shader optimizer from trying to optimize the shader again.